+++ R/tidy_add_pairwise_contrasts.R
@@ -68,6 +68,13 @@ tidy_add_pairwise_contrasts <- function(
 
   .attributes <- .save_attributes(x)
 
+  if (isTRUE(.attributes$coefficients_type == "average_marginal_effects"))
+    cli::cli_abort("Pairwise contrasts are not compatible with Average Marginal Effects.") # nolint
+  if (isTRUE(.attributes$coefficients_type == "marginal_effects"))
+    cli::cli_abort("Pairwise contrasts are not compatible with Marginal Effects.") # nolint
+  if (isTRUE(.attributes$coefficients_type == "conditional_effects"))
+    cli::cli_abort("Pairwise contrasts are not compatible with Conditional Effects.") # nolint
+
   if (is.null(conf.level))
     conf.level <- .attributes$conf.level