+++ NEWS.md
@@ -8,6 +8,9 @@
   to compute pairwise contrasts of categorical variables with `emmeans`,
   and corresponding new arguments in `tidy_plus_plus()` (#192)
 - `tidy_select_variables()` now sorts the variables according to `include` (#183)
+- New tidier `tidy_margins()` to display Average Marginal Effects (#195)
+- New tidier `tidy_all_effects()` to display Marginal Effects (#195)
+- New tidier `tidy_ggpredict()` to display Conditional Effects (#195)
 
 **New supported models**